requirements.txt
boto3==1.18.64
botocore==1.21.64
google-cloud-vision==2.5.0
pandas==1.3.3
regex==2021.10.8